using docker to run API tests (#42)

+++ src/api/users.rs
@@ -9,7 +9,11 @@ use sqlx::types::Uuid;
 
 use crate::{
     application::{
-        repository::user_repo, security::jwt_claims::AccessClaims, shared::state::SharedState,
+        api_error::ApiError,
+        api_version::{self, ApiVersion},
+        repository::user_repo,
+        security::jwt_claims::{AccessClaims, ClaimsMethods},
+        state::SharedState,
     },
     domain::models::user::User,
 };
@@ -24,72 +28,83 @@ pub fn routes() -> Router<SharedState> {
 }
 
 async fn list_users_handler(
+    api_version: ApiVersion,
     access_claims: AccessClaims,
     State(state): State<SharedState>,
-) -> Result<Json<Vec<User>>, impl IntoResponse> {
-    // ToDo: check access control
+) -> Result<Json<Vec<User>>, ApiError> {
+    tracing::trace!("api version: {}", api_version);
     tracing::trace!("authentication details: {:#?}", access_claims);
+    access_claims.validate_role_admin()?;
     match user_repo::all_users(&state).await {
         Some(users) => Ok(Json(users)),
-        None => Err(StatusCode::NOT_FOUND),
+        None => Err(StatusCode::NOT_FOUND.into()),
     }
 }
 
 async fn add_user_handler(
+    api_version: ApiVersion,
     access_claims: AccessClaims,
     State(state): State<SharedState>,
     Json(user): Json<User>,
-) -> impl IntoResponse {
+) -> Result<impl IntoResponse, ApiError> {
+    tracing::trace!("api version: {}", api_version);
     tracing::trace!("authentication details: {:#?}", access_claims);
+    access_claims.validate_role_admin()?;
     match user_repo::add_user(user, &state).await {
-        Some(user) => (StatusCode::CREATED, Json(user)).into_response(),
-        None => StatusCode::INTERNAL_SERVER_ERROR.into_response(),
+        Some(user) => Ok((StatusCode::CREATED, Json(user))),
+        None => Err(StatusCode::INTERNAL_SERVER_ERROR.into()),
     }
 }
 
 async fn get_user_handler(
     access_claims: AccessClaims,
-    Path(id): Path<Uuid>,
+    Path((version, id)): Path<(String, Uuid)>,
     State(state): State<SharedState>,
-) -> Result<Json<User>, impl IntoResponse> {
-    // ToDo: check access control
-    tracing::trace!("id: {}", id);
+) -> Result<Json<User>, ApiError> {
+    let api_version: ApiVersion = api_version::parse_version(&version)?;
+    tracing::trace!("api version: {}", api_version);
     tracing::trace!("authentication details: {:#?}", access_claims);
+    tracing::trace!("id: {}", id);
+    access_claims.validate_role_admin()?;
     match user_repo::get_user(id, &state).await {
         Some(user) => Ok(Json(user)),
-        None => Err(StatusCode::NOT_FOUND),
+        None => Err(StatusCode::NOT_FOUND.into()),
     }
 }
 
 async fn update_user_handler(
     access_claims: AccessClaims,
-    Path(id): Path<Uuid>,
+    Path((version, id)): Path<(String, Uuid)>,
     State(state): State<SharedState>,
     Json(user): Json<User>,
-) -> Result<Json<User>, impl IntoResponse> {
-    // ToDo: check access control
-    tracing::trace!("id: {}", id);
+) -> Result<Json<User>, ApiError> {
+    let api_version: ApiVersion = api_version::parse_version(&version)?;
+    tracing::trace!("api version: {}", api_version);
     tracing::trace!("authentication details: {:#?}", access_claims);
+    tracing::trace!("id: {}", id);
+    access_claims.validate_role_admin()?;
     match user_repo::update_user(id, user, &state).await {
         Some(user) => Ok(Json(user)),
-        None => Err(StatusCode::NOT_FOUND),
+        None => Err(StatusCode::NOT_FOUND.into()),
     }
 }
 
 async fn delete_user_handler(
     access_claims: AccessClaims,
-    Path(id): Path<Uuid>,
+    Path((version, id)): Path<(String, Uuid)>,
     State(state): State<SharedState>,
-) -> impl IntoResponse {
-    // ToDo: check access control
-    tracing::trace!("id: {}", id);
+) -> Result<impl IntoResponse, ApiError> {
+    let api_version: ApiVersion = api_version::parse_version(&version)?;
+    tracing::trace!("api version: {}", api_version);
     tracing::trace!("authentication details: {:#?}", access_claims);
+    tracing::trace!("id: {}", id);
+    access_claims.validate_role_admin()?;
     match user_repo::delete_user(id, &state).await {
-        Some(true) => StatusCode::OK,
-        Some(false) => {
-            tracing::warn!("User not found for deletion: {}", id);
-            StatusCode::NOT_FOUND
-        }
-        None => StatusCode::INTERNAL_SERVER_ERROR,
+        Some(true) => Ok(StatusCode::OK),
+        Some(false) => Err(ApiError {
+            status_code: StatusCode::NOT_FOUND,
+            error_message: format!("User not found for deletion: {}", id),
+        }),
+        None => Err(StatusCode::INTERNAL_SERVER_ERROR.into()),
     }
 }
